requirements.txt
Flask==2.3.3
pandas==2.3.0
numpy==2.3.0
openpyxl==3.1.2
gunicorn==21.2.0
supabase==2.3.4
requests==2.31.0
psycopg2-binary==2.9.10
